[REF] crm_lead_code: fix lints
