+++ crm_lead_code/models/crm_lead.py
@@ -23,7 +23,7 @@ class CrmLead(models.Model):
             vals['code'] = self.env['ir.sequence'].get('crm.lead')
         return super(CrmLead, self).create(vals)
 
-    @api.one
+    @api.multi
     def copy(self, default=None):
         if default is None:
             default = {}
